refactor: organize client/src into feature-based subdirectories
Move 49 files from flat client/src/ into 7 cohesive subdirectories
based on import-graph analysis: terminal/ (19), inspector/ (4),
settings/ (5), ui/ (5), rpc/ (5), input/ (5), sidebar/ (5).

All imports rewritten to use direct paths (no barrel re-exports).
Test files co-located with their source. No logic changes.

Closes #485
