+++ client/src/App.tsx
@@ -12,33 +12,33 @@ import { Title } from "@solidjs/meta";
 import { Toaster } from "solid-sonner";
 import Header from "./Header";
 import PwaInstallBar from "./PwaInstallBar";
-import Sidebar from "./Sidebar";
-import TerminalPane from "./TerminalPane";
+import Sidebar from "./sidebar/Sidebar";
+import TerminalPane from "./terminal/TerminalPane";
 import MobileKeyBar from "./MobileKeyBar";
 import CommandPalette from "./CommandPalette";
 import ShortcutsHelp from "./ShortcutsHelp";
 import ClaudeTranscriptDialog from "./ClaudeTranscriptDialog";
-import ModalDialog, { refocusTerminal } from "./ModalDialog";
+import ModalDialog, { refocusTerminal } from "./ui/ModalDialog";
 import Dialog from "@corvu/dialog";
 import Resizable from "@corvu/resizable";
 import EmptyState from "./EmptyState";
-import RightPanel from "./RightPanel";
+import RightPanel from "./inspector/RightPanel";
 import CloseConfirm, { type CloseConfirmTarget } from "./CloseConfirm";
 import { createCommands } from "./commands";
 import { exportSessionAsPdf } from "./exportSessionAsPdf";
 
 import type { TerminalId } from "kolu-common";
-import { client, wsStatus, serverProcessId } from "./rpc";
-import TransportOverlay from "./TransportOverlay";
-import { useTerminals } from "./useTerminals";
-import { useServerState } from "./useServerState";
+import { client, wsStatus, serverProcessId } from "./rpc/rpc";
+import TransportOverlay from "./rpc/TransportOverlay";
+import { useTerminals } from "./terminal/useTerminals";
+import { useServerState } from "./settings/useServerState";
 import { useThemeManager } from "./useThemeManager";
-import { useSidebar } from "./useSidebar";
-import { useShortcuts } from "./useShortcuts";
-import { useSubPanel } from "./useSubPanel";
-import { useRightPanel } from "./useRightPanel";
-import { useColorScheme } from "./useColorScheme";
-import { useTips } from "./useTips";
+import { useSidebar } from "./sidebar/useSidebar";
+import { useShortcuts } from "./input/useShortcuts";
+import { useSubPanel } from "./terminal/useSubPanel";
+import { useRightPanel } from "./inspector/useRightPanel";
+import { useColorScheme } from "./settings/useColorScheme";
+import { useTips } from "./settings/useTips";
 
 const App: Component = () => {
   const { preferences, updatePreferences } = useServerState();
