+++ client/src/inspector/RightPanel.tsx
