+++ client/src/inspector/useRightPanel.ts
