+++ client/src/rpc/streamCleanup.ts
