+++ client/src/sidebar/ActivityGraph.tsx
