+++ client/src/sidebar/ChecksIndicator.tsx
