+++ client/src/terminal/Terminal.tsx
@@ -27,17 +27,17 @@ import { ImageAddon } from "@xterm/addon-image";
 import { SerializeAddon } from "@xterm/addon-serialize";
 import "@xterm/xterm/css/xterm.css";
 import { DEFAULT_SCROLLBACK } from "kolu-common/config";
-import { FONT_FAMILY } from "./theme";
-import { client, stream } from "./rpc";
-import { isExpectedCleanupError } from "./streamCleanup";
-import { matchesAnyShortcut } from "./keyboard";
+import { FONT_FAMILY } from "../theme";
+import { client, stream } from "../rpc/rpc";
+import { isExpectedCleanupError } from "../rpc/streamCleanup";
+import { matchesAnyShortcut } from "../input/keyboard";
 import type { TerminalId } from "kolu-common";
 import SearchBar from "./SearchBar";
 import ScrollToBottom from "./ScrollToBottom";
-import { createZoom } from "./zoom";
-import { createScrollLock } from "./scrollLock";
-import { refitOnTabVisible } from "./refitOnTabVisible";
-import { viewportDimensions, setViewportDimensions } from "./useViewport";
+import { createZoom } from "../input/zoom";
+import { createScrollLock } from "../scrollLock";
+import { refitOnTabVisible } from "../refitOnTabVisible";
+import { viewportDimensions, setViewportDimensions } from "../useViewport";
 import { registerTerminalRefs, unregisterTerminalRefs } from "./terminalRefs";
 
 export type RendererType = "webgl" | "canvas";
