+++ client/src/terminal/TerminalPane.tsx
