+++ client/src/terminal/terminalDisplay.test.ts
