+++ client/src/terminal/useActivityAlerts.ts
