+++ client/src/terminal/useSubPanel.ts
